# Conflicts:
#	src/main/webapp/WEB-INF/web.xml
#	target/m2e-wtp/web-resources/META-INF/maven/in.ac.iiitkota.developers/ERP/pom.properties

+++ sources/main/postgreSQLDatabase/chats/Message.java
@@ -6,22 +6,22 @@ package postgreSQLDatabase.chats;
 import java.sql.Date;
 
 /**
- * @author Shubhi
+ * @author Megha
  *
  */
 public class Message {
 private String username;
-private int id;
-private int author;
+private long id;
+private long author;
 private String text;
 private Date time_stamp;
-public int getId() {
+public long getId() {
 	return id;
 }
-public void setId(int id) {
+public void setId(long id) {
 	this.id = id;
 }
-public int getAuthor() {
+public long getAuthor() {
 	return author;
 }
 public void setAuthor(int author) {
